:bug: Passing non-YAML files with --rules option. (#73)
The addon is passing non-YAML files in --rules option to the analyzer.
For example: when custom targets built with uploaded windup.xml files.

This does not seem to hurt anything. The analyzer output mentions that
it cannot be parsed as YAML but does not error. Fixing will mainly
avoid: confusion; red-herrings while troubleshooting.

Signed-off-by: Jeff Ortel <jortel@redhat.com>